+++ packages/cloud_firestore/cloud_firestore/lib/cloud_firestore.dart
@@ -37,7 +37,11 @@ export 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
         FieldOverrideIndex,
         Order,
         ArrayConfig,
-        QueryScope;
+        QueryScope,
+        LoadBundleTaskState,
+        average,
+        count,
+        sum;
 export 'package:firebase_core_platform_interface/firebase_core_platform_interface.dart'
     show FirebaseException;