+++ packages/firebase_crashlytics/firebase_crashlytics/lib/firebase_crashlytics.dart
@@ -6,7 +6,7 @@
 library firebase_crashlytics;
 
 import 'package:flutter/foundation.dart'
-    show kDebugMode, FlutterErrorDetails, FlutterError;
+    show DiagnosticLevel, FlutterError, FlutterErrorDetails, kDebugMode;
 
 import 'package:firebase_crashlytics_platform_interface/firebase_crashlytics_platform_interface.dart';
 import 'package:firebase_core/firebase_core.dart';